Cuando el id de cliente tiene un valor demasiado alto se sale de la capacidad del tipo int.

+++ Conciliacion/App_Web/Sitio Conicliacion 07-07-2017/SitioConciliacion/ControlesUsuario/BuscadorClienteFactura/wucBuscaClientesFacturas.ascx.cs
@@ -117,7 +117,7 @@ public partial class ControlesUsuario_BuscadorClienteFactura_wucBuscaClientesFac
                 Conexion conexion = new Conexion();
                 conexion.AbrirConexion(true);
 
-                dtClientePedidos = cliente.ObtienePedidosCliente(Convert.ToInt32(NumeroClienteFiltrar), conexion);
+                dtClientePedidos = cliente.ObtienePedidosCliente(Convert.ToInt64(NumeroClienteFiltrar), conexion);
             }
 
             return dtClientePedidos;
